pyproject.toml
[tool.poetry]
name = "accessiblegpt"
version = "0.1.0"
description = ""
authors = ["ashishdatta <ashishbdatta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.27.2"
replicate = "^0.15.4"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
